chore: remove next-themes dependency and update related configurations
- Removed the next-themes package from package.json and pnpm-lock.yaml to streamline dependencies.
- Updated tailwind.config.ts to clarify dark mode settings.
- Adjusted layout.tsx to suppress hydration warnings for better performance.
- Refactored dashboard components to enhance structure and improve loading states in settings-view.

All pre-commit checks pass: format, lint, type-check, test, build.
