+++ app/layout.tsx
@@ -40,7 +40,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
+    <html lang="en" suppressHydrationWarning>
       <head />
       <body
         className={`${inter.className} h-full overflow-x-hidden bg-background text-foreground`}
