+++ features/dashboard/components/settings-view.tsx
@@ -10,6 +10,8 @@ import {
 } from '@/components/ui/card';
 import { Label } from '@/components/ui/label';
 import { User, LogOut, Shield, Smartphone } from 'lucide-react';
+import { getBuildInfo } from '@/lib/utils/app-version';
+import { LoadingSpinner } from '@/components/ui/loading-states';
 
 interface SettingsViewProps {
   user?: any;
@@ -22,6 +24,7 @@ export function SettingsView({
   isLoggingOut,
   handleLogout,
 }: SettingsViewProps) {
+  const buildInfo = getBuildInfo();
   return (
     <div className="space-y-4 p-4">
       {/* Account Information */}
@@ -38,16 +41,28 @@ export function SettingsView({
         <CardContent className="space-y-4">
           <div className="space-y-2">
             <Label>Email</Label>
-            <p className="text-sm text-foreground">
-              {user?.email || 'Loading...'}
+            <p className="text-sm text-foreground flex items-center gap-2">
+              {user?.email ? (
+                user.email
+              ) : (
+                <>
+                  <LoadingSpinner size="sm" />
+                  Loading...
+                </>
+              )}
             </p>
           </div>
           <div className="space-y-2">
             <Label>Member Since</Label>
-            <p className="text-sm text-muted-foreground">
-              {user?.createdAt
-                ? new Date(user.createdAt).toLocaleDateString()
-                : 'Loading...'}
+            <p className="text-sm text-muted-foreground flex items-center gap-2">
+              {user?.createdAt ? (
+                new Date(user.createdAt).toLocaleDateString()
+              ) : (
+                <>
+                  <LoadingSpinner size="sm" />
+                  Loading...
+                </>
+              )}
             </p>
           </div>
         </CardContent>
@@ -99,15 +114,21 @@ export function SettingsView({
         <CardContent className="space-y-3">
           <div className="flex justify-between items-center">
             <Label className="text-sm">Version</Label>
-            <span className="text-sm text-muted-foreground">1.0.0</span>
+            <span className="text-sm text-muted-foreground">
+              {buildInfo.version}
+            </span>
           </div>
           <div className="flex justify-between items-center">
             <Label className="text-sm">Build</Label>
-            <span className="text-sm text-muted-foreground">Production</span>
+            <span className="text-sm text-muted-foreground">
+              {buildInfo.build}
+            </span>
           </div>
           <div className="flex justify-between items-center">
             <Label className="text-sm">Platform</Label>
-            <span className="text-sm text-muted-foreground">PWA</span>
+            <span className="text-sm text-muted-foreground">
+              {buildInfo.platform}
+            </span>
           </div>
         </CardContent>
       </Card>
@@ -128,8 +149,17 @@ export function SettingsView({
             variant="outline"
             className="w-full sm:w-auto"
           >
-            <LogOut className="h-4 w-4 mr-2" />
-            {isLoggingOut ? 'Logging out...' : 'Logout'}
+            {isLoggingOut ? (
+              <>
+                <LoadingSpinner size="sm" className="mr-2" />
+                Logging out...
+              </>
+            ) : (
+              <>
+                <LogOut className="h-4 w-4 mr-2" />
+                Logout
+              </>
+            )}
           </Button>
         </CardContent>
       </Card>
