+++ package.json
@@ -76,7 +76,6 @@
     "isomorphic-dompurify": "^2.26.0",
     "lucide-react": "^0.454.0",
     "next": "15.2.4",
-    "next-themes": "^0.4.4",
     "openai": "^5.10.1",
     "react": "^19",
     "react-day-picker": "8.10.1",
