+++ tailwind.config.ts
@@ -3,7 +3,7 @@ import type { Config } from 'tailwindcss';
 // all in fixtures is set to tailwind v3 as interims solutions
 
 const config: Config = {
-  darkMode: ['class'],
+  // Dark mode only - no class switching needed
   content: [
     './pages/**/*.{js,ts,jsx,tsx,mdx}',
     './components/**/*.{js,ts,jsx,tsx,mdx}',
